[#2127] Extract reusable PendingSaving hook
